ai-service/requirements.txt
fastapi
uvicorn
konlpy
openai-whisper
python-multipart
librosa
numpy
requests